feat - проверка на уровне json-схемы для EpicHandler. Настроил возврат сообщений пользователю в виде json

+++ src/main/java/http/server/handler/EpicHandler.java
@@ -19,9 +19,11 @@ import java.nio.charset.StandardCharsets;
 import java.time.Instant;
 import java.util.OptionalInt;
 
-public class EpicHandler implements HttpHandler {
-
+public class EpicHandler extends HandlerHelper implements HttpHandler {
     private static final Charset DEFAULT_CHARSET = StandardCharsets.UTF_8;
+    private static final String CREATE_SCHEMA = "/create_epic_schema.json";
+    private static final String UPDATE_SCHEMA = "/update_epic_schema.json";
+    String[] paths = {"^/tasks/epic/\\?id=\\d+$", "^/tasks/epic/?$"};
     private final TaskManager taskManager;
     private final Gson gson;
 
@@ -42,25 +44,44 @@ public class EpicHandler implements HttpHandler {
         InputStream inputStream = exchange.getRequestBody();
         String requestBody = new String(inputStream.readAllBytes(), DEFAULT_CHARSET);
 
-        switch (method) {
-            case "POST":
-                response = createEpic(requestBody);
-                break;
-            case "PUT":
-                response = updateEpic(requestBody);
-                break;
-            case "GET":
-                response = getEpic(getIdFromPath(exchange));
-                break;
-            case "DELETE":
-                if (isTotalDelete(exchange)) {
-                    response = deleteAllEpics();
-                } else {
-                    response = deleteEpic(getIdFromPath(exchange));
-                }
-                break;
-            default:
-                response = new Response(405, "Метод не поддерживается. Доступны: GET, POST, DELETE, PUT");
+        if (isCorrectPath(exchange, paths)) {
+            switch (method) {
+                case "POST":
+                    if (validateJson(requestBody, CREATE_SCHEMA)) {
+                        response = createEpic(requestBody);
+                    } else {
+                        response = new Response(400, getJsonError(requestBody, CREATE_SCHEMA));
+                    }
+                    break;
+                case "PUT":
+                    if (validateJson(requestBody, UPDATE_SCHEMA)) {
+                        response = updateEpic(requestBody);
+                    } else {
+                        response = new Response(400, getJsonError(requestBody, UPDATE_SCHEMA));
+                    }
+                    break;
+                case "GET":
+                    if (exchange.getRequestURI().getPath() != null) {
+                        response = getEpic(getIdFromPath(exchange));
+                    } else {
+                        response = new Response(400, gson.toJson(constructResponse(false, 400,
+                                "Данный эндпоинт не реализован.")));
+                    }
+                    break;
+                case "DELETE":
+                    if (isTotalDelete(exchange)) {
+                        response = deleteAllEpics();
+                    } else {
+                        response = deleteEpic(getIdFromPath(exchange));
+                    }
+                    break;
+                default:
+                    response = new Response(405, gson.toJson(constructResponse(false, 405,
+                            "Метод не поддерживается. Доступны: GET, POST, DELETE, PUT")));
+            }
+        } else {
+            response = new Response(405, gson.toJson(constructResponse(false, 405,
+                    "Данный эндпоинт не реализован.")));
         }
 
         Headers headers2 = exchange.getResponseHeaders();
@@ -74,19 +95,6 @@ public class EpicHandler implements HttpHandler {
         }
     }
 
-    private boolean isTotalDelete(HttpExchange exchange) {
-        return exchange.getRequestURI().getQuery() == null;
-    }
-
-    private int getIdFromPath(HttpExchange exchange) {
-        return Integer.parseInt(exchange.getRequestURI().getQuery().split("=")[1]);
-    }
-
-
-    private boolean isUpdate(String requestBody) {
-        return requestBody.contains("\"id\": ");
-    }
-
     private Epic getCreatedEpic() {
         OptionalInt newEpicId = taskManager.getEpics().stream()
                 .map(Task::getId)
@@ -96,14 +104,10 @@ public class EpicHandler implements HttpHandler {
     }
 
     private Response createEpic(String requestBody) {
-        if (requestBody.isEmpty()) {
-            return new Response(400, "Тело запроса не должно быть пустым.");
-        } else {
             Epic epicData = gson.fromJson(requestBody, Epic.class);
             taskManager.createEpic(epicData);
             Epic createdEpic = getCreatedEpic();
             return new Response(201, gson.toJson(createdEpic));
-        }
     }
 
     private Response updateEpic(String requestBody) {
@@ -111,10 +115,9 @@ public class EpicHandler implements HttpHandler {
             Epic epic  = gson.fromJson(requestBody, Epic.class);
             int epicId = epic.getId();
             taskManager.updateEpic(epicId, epic);
-
             return new Response(200, gson.toJson(taskManager.getEpic(epicId)));
         } catch (ManagerValidateException e) {
-            return new Response(404, e.getMessage());
+            return new Response(404, gson.toJson(constructResponse(false, 405, e.getMessage())));
         }
     }
 
@@ -122,7 +125,8 @@ public class EpicHandler implements HttpHandler {
         Epic epic = taskManager.getEpicById(id);
 
         if (epic == null) {
-            return new Response(404, String.format("Эпик с id = %d не найден.", id));
+            return new Response(404, gson.toJson(constructResponse(
+                    false, 404, String.format("Эпик с id = %d не найден.", id))));
         } else {
             return new Response(200, gson.toJson(epic));
         }
@@ -132,7 +136,8 @@ public class EpicHandler implements HttpHandler {
         Epic epic = taskManager.getEpic(id);
 
         if (epic == null) {
-            return new Response(400, "Эпик, который вы хотите удалить не существует.");
+            return new Response(404, gson.toJson(constructResponse(
+                    false, 404, String.format("Эпик с id = %d не найден.", id))));
         } else {
             taskManager.removeEpicById(id);
             return new Response(200, String.format("Эпик с id = %d удален.", id));
@@ -141,6 +146,7 @@ public class EpicHandler implements HttpHandler {
 
     private Response deleteAllEpics() {
         taskManager.removeAllEpics();
-        return new Response(200, "Все эпики удалены.");
+        return new Response(200, gson.toJson(constructResponse(
+                true, 200, "Все задачи удалены.")));
     }
 }
